refatcor: Move XmlParserTests to CompletionEngineTests.Parsing
